+++ tests/extendrtests/src/rust/src/lib.rs
@@ -1,4 +1,4 @@
-use extendr_api::*;
+use extendr_api::prelude::*;
 
 /// Return string `"Hello world!"` to R.
 /// @export
